requirements.txt
streamlit
librosa
numpy
scikit-learn
xgboost
matplotlib
soundfile
joblib
opencv-python
pillow
torch
torchvision
fastapi
uvicorn